feat: move old requests_html into package
split functionalities into separated modules
