+++ requests_html/py.typed
@@ -1,0 +1,1 @@
+
