+++ requests_html/sessions.py
@@ -1,0 +1,102 @@
+import asyncio
+from concurrent.futures import ThreadPoolExecutor
+from functools import partial
+
+from requests_html.html_response import HTMLResponse
+from requests_html.globals import DEFAULT_ENCODING
+import pyppeteer
+import requests
+
+from requests_html import user_agent
+
+
+class BaseSession(requests.Session):
+    """ A consumable session, for cookie persistence and connection pooling,
+    amongst other things.
+    """
+
+    def __init__(self, mock_browser : bool = True, verify : bool = True,
+                 browser_args : list = ['--no-sandbox']):
+        super().__init__()
+
+        # Mock a web browser's user agent.
+        if mock_browser:
+            self.headers['User-Agent'] = user_agent()
+
+        self.hooks['response'].append(self.response_hook)
+        self.verify = verify
+
+        self.__browser_args = browser_args
+
+
+    def response_hook(self, response, **kwargs) -> HTMLResponse:
+        """ Change response encoding and replace it by a HTMLResponse. """
+        if not response.encoding:
+            response.encoding = DEFAULT_ENCODING
+        return HTMLResponse._from_response(response, self)
+
+    @property
+    async def browser(self):
+        if not hasattr(self, "_browser"):
+            self._browser = await pyppeteer.launch(ignoreHTTPSErrors=not(self.verify), headless=True, args=self.__browser_args)
+
+        return self._browser
+
+
+class HTMLSession(BaseSession):
+
+    def __init__(self, **kwargs):
+        super(HTMLSession, self).__init__(**kwargs)
+
+    @property
+    def browser(self):
+        if not hasattr(self, "_browser"):
+            self.loop = asyncio.get_event_loop()
+            if self.loop.is_running():
+                raise RuntimeError("Cannot use HTMLSession within an existing event loop. Use AsyncHTMLSession instead.")
+            self._browser = self.loop.run_until_complete(super().browser)
+        return self._browser
+
+    def close(self):
+        """ If a browser was created close it first. """
+        if hasattr(self, "_browser"):
+            self.loop.run_until_complete(self._browser.close())
+        super().close()
+
+
+class AsyncHTMLSession(BaseSession):
+    """ An async consumable session. """
+
+    def __init__(self, loop=None, workers=None,
+                 mock_browser: bool = True, *args, **kwargs):
+        """ Set or create an event loop and a thread pool.
+
+            :param loop: Asyncio loop to use.
+            :param workers: Amount of threads to use for executing async calls.
+                If not pass it will default to the number of processors on the
+                machine, multiplied by 5. """
+        super().__init__(*args, **kwargs)
+
+        self.loop = loop or asyncio.get_event_loop()
+        self.thread_pool = ThreadPoolExecutor(max_workers=workers)
+
+    def request(self, *args, **kwargs):
+        """ Partial original request func and run it in a thread. """
+        func = partial(super().request, *args, **kwargs)
+        return self.loop.run_in_executor(self.thread_pool, func)
+
+    async def close(self):
+        """ If a browser was created close it first. """
+        if hasattr(self, "_browser"):
+            await self._browser.close()
+        super().close()
+
+    def run(self, *coros):
+        """ Pass in all the coroutines you want to run, it will wrap each one
+            in a task, run it and wait for the result. Return a list with all
+            results, this is returned in the same order coros are passed in. """
+        tasks = [
+            asyncio.ensure_future(coro()) for coro in coros
+        ]
+        done, _ = self.loop.run_until_complete(asyncio.wait(tasks))
+        return [t.result() for t in done]
