+++ requests_html/typing.py
@@ -1,0 +1,24 @@
+from typing import Set, Union, List, MutableMapping
+from lxml.html import HtmlElement
+from parse import Result
+
+# Typing.
+_Find = Union[List['Element'], 'Element']
+_XPath = Union[List[str], List['Element'], str, 'Element']
+_Result = Union[List['Result'], 'Result']
+_HTML = Union[str, bytes]
+_BaseHTML = str
+_UserAgent = str
+_DefaultEncoding = str
+_URL = str
+_RawHTML = bytes
+_Encoding = str
+_LXML = HtmlElement
+_Text = str
+_Search = Result
+_Containing = Union[str, List[str]]
+_Links = Set[str]
+_Attrs = MutableMapping
+_Next = Union['HTML', List[str]]
+_NextSymbol = List[str]
+
